requirements.txt
discord.py>=2.0.0
yt-dlp>=2023.1.1
openai-whisper>=20230124
python-dotenv>=1.0.0
google-generativeai>=0.3.0
ffmpeg-python>=0.2.0